libsdhcdrivers: Change block size type to size_t
Block size shall be of `size_t` type due to the following reasons:

1. Block size cannot be bigger than the addressable memory so it is best
   described with the `size_t`.
2. Negative value of block size is not acceptable.
3. More compatibility with the std e.g. when memseting block.

+++ libsdhcdrivers/include/sdhc/mmc.h
@@ -21,7 +21,7 @@ typedef struct mmc_card *mmc_card_t;
 typedef void (*mmc_cb)(mmc_card_t mmc_card, int status, size_t bytes_transferred, void *token);
 
 
-static inline int mmc_block_size(mmc_card_t mmc_card)
+static inline size_t mmc_block_size(mmc_card_t mmc_card)
 {
     return 512;
 }
